+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [latest]
+### Added
+- colours in the geometry image can now be changed [#960](https://github.com/spatial-model-editor/spatial-model-editor/issues/960)
+- voxels can be selected in the 3d rendering of the geometry image [#979](https://github.com/spatial-model-editor/spatial-model-editor/issues/979)
+
 ## [1.6.0] - 2024-03-26
 ### Added
 - support for using the DuneCopasi simulator with 3d models [#937](https://github.com/spatial-model-editor/spatial-model-editor/issues/937)